+++ src/__tests__/resource-context.test.ts
@@ -1,5 +1,6 @@
 import ResourceContext from '../resource-context';
-import { create } from '../allocation';
+import { create } from '../';
+import { createContext } from '../inherited-context';
 
 describe('ResourceContext', () => {
   async function Test() {
@@ -71,5 +72,115 @@ describe('ResourceContext', () => {
     }
 
     await expect(create(Allocator)).rejects.toThrow(/already destroyed/i);
+  });
+
+  it('can set and retrieve context', async () => {
+    const SharedValue = createContext(() => 'none');
+
+    const Test = async (resource: ResourceContext) => {
+      expect(resource.getContext(SharedValue)).toBe('none');
+
+      resource.setContext(SharedValue, 'saved');
+      expect(resource.getContext(SharedValue)).toBe('saved');
+
+      resource.setContext(SharedValue, 'updated');
+      expect(resource.getContext(SharedValue)).toBe('updated');
+
+      return { value: [] };
+    };
+
+    await expect(create(Test)).resolves.toEqual([]);
+  });
+
+  it('passes context to child resources', async () => {
+    const SharedValue = createContext<null | string>(() => null);
+
+    const Child = async (ctx: ResourceContext) => ({
+      value: { content: ctx.getContext(SharedValue) },
+    });
+
+    const Parent = async (resource: ResourceContext) => {
+      resource.setContext(SharedValue, 'inherited');
+      return { value: await resource.create(Child) };
+    };
+
+    await expect(create(Parent)).resolves.toEqual({ content: 'inherited' });
+  });
+
+  it('can override context without affecting the parent', async () => {
+    const Message = createContext<null | string>(() => null);
+    const Child = async (resource: ResourceContext) => {
+      // This should *NOT* affect the parent context.
+      resource.setContext(Message, 'child context');
+
+      return {
+        value: { content: resource.getContext(Message) },
+      };
+    };
+
+    const Parent = async (resource: ResourceContext) => {
+      resource.setContext(Message, 'parent context');
+      const child = await resource.create(Child);
+
+      expect(resource.getContext(Message)).toBe('parent context');
+
+      return {
+        value: {
+          content: resource.getContext(Message),
+          child,
+        },
+      };
+    };
+
+    await expect(create(Parent)).resolves.toEqual({
+      content: 'parent context',
+      child: { content: 'child context' },
+    });
+  });
+
+  it('allows two siblings to have different context values', async () => {
+    const Message = createContext<null | string>(() => null);
+    const Child = async (resource: ResourceContext, msg: string) => {
+      resource.setContext(Message, msg);
+
+      return {
+        value: { getMessage: () => resource.getContext(Message) },
+      };
+    };
+
+    const Parent = async (resource: ResourceContext) => {
+      resource.setContext(Message, 'parent context');
+
+      return {
+        value: await Promise.all([
+          resource.create(Child, 'child context 1'),
+          resource.create(Child, 'child context 2'),
+        ]),
+      };
+    };
+
+    const [child1, child2] = await create(Parent);
+    expect(child1.getMessage()).toBe('child context 1');
+    expect(child2.getMessage()).toBe('child context 2');
+  });
+
+  it('provides a live view of the current value, not just a snapshot', async () => {
+    const Message = createContext(() => 'default');
+    const Child = async (resource: ResourceContext) => ({
+      value: { getMessage: () => resource.getContext(Message) },
+    });
+
+    const Parent = async (resource: ResourceContext) => ({
+      value: {
+        setMessage: (msg: string) => resource.setContext(Message, msg),
+        child: await resource.create(Child),
+      },
+    });
+
+    const { setMessage, child } = await create(Parent);
+    expect(child.getMessage()).toBe('default');
+
+    setMessage('updated');
+    expect(child.getMessage()).toBe('updated');
   });
 });