+++ src/index.ts
@@ -1,4 +1,5 @@
 export { default as bindContext } from './bind-context';
-export { create, destroy } from './allocation';
+export { create, destroy } from './roots';
+export { createContext } from './inherited-context';
 export type { Resource, ResourceHandle } from './types';
 export type { default as ResourceContext } from './resource-context';